+++ projects/Week12-Completed/src/Graphics/UniformBuffer.cpp
@@ -5,7 +5,7 @@ AbstractUniformBuffer::~AbstractUniformBuffer() {
 	delete[] _rawData;
 }
 
-AbstractUniformBuffer::AbstractUniformBuffer(size_t sizeInBytes, BufferUsage usage /*= BufferUsage::DynamicDraw*/) :
+AbstractUniformBuffer::AbstractUniformBuffer(uint32_t sizeInBytes, BufferUsage usage /*= BufferUsage::DynamicDraw*/) :
 	IBuffer(BufferType::Uniform, usage),
 	_rawData(nullptr)
 {
@@ -15,7 +15,7 @@ AbstractUniformBuffer::AbstractUniformBuffer(size_t sizeInBytes, BufferUsage usa
 	glNamedBufferData(_rendererId, _size, _rawData, (GLenum)_usage);
 }
 
-void AbstractUniformBuffer::LoadData(const void* data, size_t elementSize, size_t elementCount) {
+void AbstractUniformBuffer::LoadData(const void* data, uint32_t elementSize, uint32_t elementCount) {
 	size_t dataSize = elementCount * elementSize;
 	LOG_ASSERT(dataSize <= _size, "Data exceeds the bounds of this UBO");
 	// Copy data from the data given to our internal buffer
